Wire in repository clone flow
